♻️  Extract functions to new file
