+++ bootstrapping_tools/__init__.py
@@ -2,3 +2,4 @@
 
 __version__ = "0.4.0"
 from .bootstrapping import *  # noqa
+from .resample_by_blocks import *  # noqa
