+++ bootstrapping_tools/bootstrapping.py
@@ -304,18 +304,3 @@ def remove_outlier(method, data, **kwargs):
     assert method in outlier_method, "No se reconoce el método de filtrado"
     data = outlier_method[method](data, **kwargs)
     return data
-
-
-def resample_data_by_blocks(original_sample, block_numbers, blocks_length):
-    n_rows_original = len(original_sample)
-    rows = get_rows(block_numbers, n_rows_original, blocks_length)
-    resample = original_sample.loc[rows, :]
-    return resample
-
-
-def get_rows(block_numbers, n_rows_data, blocks_length):
-    aux = np.arange(n_rows_data)
-    rows = []
-    for i in block_numbers:
-        rows.extend(aux[0 + i : blocks_length + i])
-    return rows
